eval 2 correction

+++ eval2/eval.go
@@ -24,7 +24,7 @@ func main() {
 	var ActionsPerTimestep int = 10   // ?
 	var SimulationDuration int = 2000 // 200,000/5 = 40,000
 
-	var currNodeCount int = int(InitialNodeCount)
+	var currNodeCount uint32 = InitialNodeCount
 
 	// Slice containing all nodes
 	var nodes []*freenet.Node
@@ -80,12 +80,12 @@ func main() {
 
 		// New node addition every 5 timesteps
 		if i%5 == 0 {
-			nodes = append(nodes, freenet.NewNode(uint32(currNodeCount)))
-			nodes[uint32(currNodeCount)].Start()
+			nodes = append(nodes, freenet.NewNode(currNodeCount))
+			nodes[currNodeCount].Start()
 
 			freenet.HopsToLiveDefault = 10
-			dstID := rand.Intn(currNodeCount)
-			nodes[uint32(currNodeCount)].SendRequestJoin(nodes[dstID])
+			dstID := rand.Intn(int(currNodeCount))
+			nodes[currNodeCount].SendRequestJoin(nodes[dstID])
 			time.Sleep(1 * time.Millisecond)
 
 			freenet.HopsToLiveDefault = 20
